refactor : ProductUiState 클래스 패키지 이동 및 로직 변경
- ProductUiState 클래스를 `woowacourse.shopping.view.main.vm`에서 `woowacourse.shopping.view.main.vm.state`로 이동했습니다.
- 수량 증가/감소 관련 로직을 ProductState로 이전하고, ProductUiState는 해당 ProductState를 교체하도록 수정했습니다.
- 장바구니에 상품 추가 가능 여부, 수량 증가 가능 여부를 확인하는 로직을 추가했습니다.
