+++ app/src/main/java/woowacourse/shopping/view/main/vm/state/ProductUiState.kt
@@ -1,0 +1,42 @@
+package woowacourse.shopping.view.main.vm.state
+
+data class ProductUiState(
+    val items: List<ProductState> = emptyList(),
+    val load: LoadState = LoadState.CannotLoad,
+) {
+    fun modifyUiState(item: ProductState): ProductUiState {
+        val targetIndex = targetIndex(item.item.id)
+        val mutableItems = items.toMutableList()
+        mutableItems[targetIndex] = item
+        return copy(items = mutableItems)
+    }
+
+    fun isAddedProduct(productId: Long): CartSavingState {
+        val targetIndex = targetIndex(productId)
+        val target = items[targetIndex]
+
+        return target.isSaveInCart()
+    }
+
+    fun canIncreaseCartQuantity(productId: Long): IncreaseState {
+        val targetIndex = targetIndex(productId)
+        val target = items[targetIndex]
+        val result = target.increase()
+
+        return result
+    }
+
+    fun itemCount() = items.size
+
+    fun addItems(
+        newItems: List<ProductState>,
+        hasNextPage: Boolean,
+    ): ProductUiState {
+        return copy(
+            items = items + newItems,
+            load = LoadState.of(hasNextPage),
+        )
+    }
+
+    private fun targetIndex(productId: Long) = items.indexOfFirst { it.item.id == productId }
+}
